flatten shm::pool into shm
This also commonizes the way in which pools are constructed
